requirement.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
black==25.1.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.13
h11==0.16.0
idna==3.10
inflect==7.5.0
isort==6.0.1
more-itertools==10.7.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlacodegen==3.0.0
SQLAlchemy==2.0.41
starlette==0.46.2
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
